backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
geopandas==0.14.1
rasterio==1.3.9
netCDF4==1.6.4
xarray==2023.11.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
beanie==1.25.0  # ODM for MongoDB 